Dissolve `enhancements` project.
