+++ codepropertygraph/src/main/scala/io/shiftleft/Implicits.scala
@@ -1,0 +1,31 @@
+package io.shiftleft
+
+object Implicits {
+
+  /**
+    * A wrapper around a Java iterator that throws a proper NoSuchElementException.
+    *
+    * Proper in this case means an exception with a stack trace.
+    * This is intended to be used as a replacement for next() on the iterators
+    * returned from TinkerPop since those are missing stack traces.
+    */
+  implicit class JavaIteratorDeco[T](iterator: java.util.Iterator[T]) {
+    def nextChecked: T = {
+      try {
+        iterator.next
+      } catch {
+        case _: NoSuchElementException =>
+          throw new NoSuchElementException()
+      }
+    }
+
+    def nextOption: Option[T] = {
+      if (iterator.hasNext) {
+        Some(iterator.next)
+      } else {
+        None
+      }
+    }
+  }
+
+}
