+++ codepropertygraph/src/main/scala/io/shiftleft/diffgraph/DiffGraphProtoSerializer.scala
