+++ codepropertygraph/src/main/scala/io/shiftleft/diffgraph/Util.scala
