+++ codepropertygraph/src/main/scala/io/shiftleft/passes/ParallelIteratorExecutor.scala
@@ -1,8 +1,8 @@
 package io.shiftleft.passes
 
-import scala.concurrent.{Await, Future}
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration.Duration
+import scala.concurrent.{Await, Future}
 
 class ParallelIteratorExecutor[T](iterator: Iterator[T]) {
   def map[D](func: T => D): Iterator[D] = {
