+++ dataflowengine/src/main/scala/io/shiftleft/passes/dataflows/steps/TrackingPoint.scala
@@ -4,9 +4,8 @@ import gremlin.scala._
 import io.shiftleft.codepropertygraph.generated._
 import io.shiftleft.queryprimitives.steps.{NodeSteps, Steps}
 import shapeless.HList
-import io.shiftleft.queryprimitives
 import io.shiftleft.queryprimitives.steps.types.structure.Method
-import io.shiftleft.queryprimitives.steps.Implicits.JavaIteratorDeco
+import io.shiftleft.Implicits.JavaIteratorDeco
 import io.shiftleft.queryprimitives.utils.{ExpandTo, MemberAccess}
 import org.apache.tinkerpop.gremlin.structure.Direction
