--- enhancements/build.sbt
@@ -1,7 +1,0 @@
-name := "enhancements"
-
-dependsOn(Projects.queryPrimitives)
-
-libraryDependencies ++= Seq(
-  "org.scalatest" %% "scalatest" % "3.0.3" % Test
-)
