+++ query-primitives/src/main/scala/io/shiftleft/queryprimitives/steps/Implicits.scala
@@ -12,30 +12,4 @@ object Implicits {
       raw.asInstanceOf[GremlinScala.Aux[NodeType, Labels]]
   }
 
-  /**
-    * A wrapper around a Java iterator that throws a proper NoSuchElementException.
-    *
-    * Proper in this case means an exception with a stack trace.
-    * This is intended to be used as a replacement for next() on the iterators
-    * returned from TinkerPop since those are missing stack traces.
-    */
-  implicit class JavaIteratorDeco[T](iterator: java.util.Iterator[T]) {
-    def nextChecked: T = {
-      try {
-        iterator.next
-      } catch {
-        case _: NoSuchElementException =>
-          throw new NoSuchElementException()
-      }
-    }
-
-    def nextOption: Option[T] = {
-      if (iterator.hasNext) {
-        Some(iterator.next)
-      } else {
-        None
-      }
-    }
-  }
-
 }
