+++ query-primitives/src/main/scala/io/shiftleft/queryprimitives/steps/nodemethods/WithinMethodMethods.scala
@@ -2,7 +2,7 @@ package io.shiftleft.queryprimitives.steps.nodemethods
 
 import io.shiftleft.codepropertygraph.generated.nodes.NodeVisitor
 import io.shiftleft.codepropertygraph.generated.{EdgeTypes, nodes}
-import io.shiftleft.queryprimitives.steps.Implicits.JavaIteratorDeco
+import io.shiftleft.Implicits.JavaIteratorDeco
 import io.shiftleft.queryprimitives.steps.visitormixins.ExpressionGeneralization
 import org.apache.tinkerpop.gremlin.structure.Direction
