+++ query-primitives/src/main/scala/io/shiftleft/queryprimitives/utils/ExpandTo.scala
@@ -3,7 +3,7 @@ package io.shiftleft.queryprimitives.utils
 import gremlin.scala.GremlinScala
 import io.shiftleft.codepropertygraph.generated._
 import org.apache.tinkerpop.gremlin.structure.{Direction, Vertex}
-import io.shiftleft.queryprimitives.steps.Implicits.JavaIteratorDeco
+import io.shiftleft.Implicits.JavaIteratorDeco
 
 import scala.annotation.tailrec
 import scala.collection.JavaConverters._
