+++ semanticcpg/src/main/scala/io/shiftleft/passes/linking/memberaccesslinker/MemberAccessLinker.scala
@@ -5,7 +5,7 @@ import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.codepropertygraph.generated.{EdgeTypes, NodeKeys, Operators, nodes}
 import io.shiftleft.diffgraph.DiffGraph
 import io.shiftleft.passes.{CpgPass, ParallelIteratorExecutor}
-import io.shiftleft.queryprimitives.steps.Implicits.JavaIteratorDeco
+import io.shiftleft.Implicits.JavaIteratorDeco
 import org.apache.logging.log4j.{LogManager, Logger}
 import org.apache.tinkerpop.gremlin.structure.Direction
